Extract simple window settings to their own component
